% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aerialPhotoFrame-package.R
\docType{package}
\name{aerialPhotoFrame-package}
\alias{aerialPhotoFrame}
\alias{aerialPhotoFrame-package}
\title{aerialPhotoFrame: Create Ground Footprint Wireframes from Aerial Photos}
\description{
Identify the ground extent (AKA wireframe) of a photo taken from the air, and create a spatial (sf) object. Useful in calculating area sampled in aerial photos and approximate spatial location. Should be applicable for photos taken from airplane or drone.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/WyoGFD/aerialPhotoFrame}
  \item Report bugs at \url{https://github.com/WyoGFD/aerialPhotoFrame/issues}
}

}
\author{
\strong{Maintainer}: Jason Carlisle \email{jason.carlisle@wyo.gov}

Authors:
\itemize{
  \item Eric Newkirk \email{eric.newkirk@wyo.gov}
}

}
\keyword{internal}
