% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAltitude.R
\name{calcAltitude}
\alias{calcAltitude}
\title{Calculate altitude above ground level}
\usage{
calcAltitude(pts, res = 1, force.redo = FALSE)
}
\arguments{
\item{pts}{sf object of camera locations with Z dimension as altitude of
camera above mean sea level in meters.}

\item{res}{character scalar indicating the resolution of digital elevation
model to use.  Passed to \code{res} argument of \code{FedData::get_ned}.
Default is \code{"1"} for 1 arc-second.}

\item{force.redo}{logical indicating whether the digital elevation model
raster should be re-downloaded to the local machine if it already exists.
Passed to \code{force.redo} argument of \code{FedData::get_ned}.
Default is \code{FALSE}.  If your output has NAs in the \code{elevAGL}
column, try setting this to \code{TRUE}.}
}
\value{
sf object, same number of rows as input, with additional columns
indicating the ground elevation (m) and altitude above ground level (AGL)
of the camera (m).
}
\description{
Calculate altitude above ground level
}
\examples{
\dontrun{
# Create points for camera locations
pts <- extractPoints("inst/extdata")
names(pts)
plot(sf::st_geometry(pts))

# Add altitude, specifically above-ground-level
pts <- calcAltitude(pts)
names(pts)
hist(pts$elevAGL, xlab = "Camera altitude AGL meters")
}
}
