% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPoints.R
\name{extractPoints}
\alias{extractPoints}
\title{Extract camera location (3D point) from photos}
\usage{
extractPoints(dir, validExt = c("JPG", "jpeg"), crs = 4326)
}
\arguments{
\item{dir}{Character, directory containing photos to process.}

\item{validExt}{Character vector, only files in \code{dir} that have an
extension specified here will be processed.  Exclude the \code{.} (dot) from
extensions (e.g., \code{JPG} instead of \code{.JPG})  Use to ignore any
non-photo files in \code{dir}. A value of \code{NULL} will read all files.}

\item{crs}{numeric scalar, EPSG code for coordinate reference system of GPS
coordinates in photo metadata.}
}
\value{
sf object of the camera location (X, Y, Z) associated with each
photo.
}
\description{
Assumes the point is the location of the camera in X, Y, Z.
}
\examples{
\dontrun{
# Create points for camera locations
pts <- extractPoints("inst/extdata")
names(pts)
plot(sf::st_geometry(pts))
}
}
