% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wireFrame.R
\name{wireFrame}
\alias{wireFrame}
\title{Create wireframe of photo footprints}
\usage{
wireFrame(pts, crs = 26912)
}
\arguments{
\item{pts}{sf object of camera locations with column indicating the camera's
altitude above ground level (m).  Output from \code{calcAltitude}.}

\item{crs}{numeric scalar, EPSG code for coordinate reference system to use
for planar calculations and for output wireframes.}
}
\value{
sf object (POLYGON) of ground area photographed for each photo,
includes same data columns as input \code{pts}, has projection specified by
input \code{crs}.
}
\description{
Create wireframe of photo footprints
}
\examples{
\dontrun{
# Create points for camera locations
pts <- extractPoints("inst/extdata")
names(pts)
plot(sf::st_geometry(pts))

# Add altitude, specifically above-ground-level
pts <- calcAltitude(pts)
names(pts)
hist(pts$elevAGL, xlab = "Camera altitude AGL")

# Create wireframe of each photo's field of view
# AKA the area of the ground photographed
wires <- wireFrame(pts, crs = 26912)
plot(sf::st_geometry(wires))
plot(sf::st_geometry(pts), col = "black", add = TRUE)
}
}
